Move page mutations
