--- saleor/dashboard/graphql/page/types.py
@@ -1,5 +1,0 @@
-from ....page import models
-
-
-def resolve_all_pages():
-    return models.Page.objects.all().distinct()
