+++ saleor/graphql/page/types.py
@@ -8,9 +8,15 @@ class Page(CountableDjangoObjectType):
     class Meta:
         description = """A static page that can be manually added by a shop
         operator through the dashboard."""
-        model = models.Page
         interfaces = [relay.Node]
+        model = models.Page
+        filter_fields = ['id', 'name']
 
 
-def resolve_pages():
+def resolve_pages(user):
+    if user.is_authenticated and user.is_active and user.is_staff:
+        return models.Page.objects.all().distinct()
     return models.Page.objects.public().distinct()
+
+def resolve_all_pages():
+    return models.Page.objects.all().distinct()
