--- tests/dashboard/test_graphql.py
@@ -1,130 +1,0 @@
-import json
-from unittest.mock import patch
-
-import graphene
-import pytest
-from django.shortcuts import reverse
-
-from saleor.dashboard.graphql.core.mutations import (
-    ModelFormMutation, ModelFormUpdateMutation)
-
-from ..utils import get_graphql_content
-
-
-@patch('saleor.dashboard.graphql.core.mutations.convert_form_fields')
-@patch('saleor.dashboard.graphql.core.mutations.convert_form_field')
-def test_model_form_mutation(
-        mocked_convert_form_field, mocked_convert_form_fields,
-        model_form_class):
-
-    mocked_convert_form_fields.return_value = {
-        model_form_class._meta.fields: mocked_convert_form_field.return_value}
-
-    class TestMutation(ModelFormMutation):
-        test_field = graphene.String()
-
-        class Arguments:
-            test_input = graphene.String()
-
-        class Meta:
-            form_class = model_form_class
-            return_field_name = 'test_return_field'
-
-    meta = TestMutation._meta
-    assert meta.form_class == model_form_class
-    assert meta.model == 'test_model'
-    assert meta.return_field_name == 'test_return_field'
-    arguments = meta.arguments
-    # check if declarative arguments are present
-    assert 'test_input' in arguments
-    # check if model form field is present
-    mocked_convert_form_fields.assert_called_with(model_form_class, None)
-    assert 'test_field' in arguments
-
-    output_fields = meta.fields
-    assert 'test_return_field' in output_fields
-    assert 'errors' in output_fields
-
-
-@patch('saleor.dashboard.graphql.core.mutations')
-def test_model_form_update_mutation(model_form_class):
-    class TestUpdateMutation(ModelFormUpdateMutation):
-        class Meta:
-            form_class = model_form_class
-            return_field_name = 'test_return_field'
-
-    meta = TestUpdateMutation._meta
-    assert 'id' in meta.arguments
-
-
-def test_page_create_mutation(admin_client):
-    query = """
-        mutation CreatePage(
-            $slug: String!,
-            $title: String!,
-            $content: String!,
-            $isVisible: Boolean!) {
-                pageCreate(slug: $slug,
-                title: $title,
-                content: $content,
-                isVisible: $isVisible) {
-                    page {
-                        id
-                        title
-                        content
-                        slug
-                        isVisible
-                      }
-                      errors {
-                        message
-                        field
-                      }
-                    }
-                  }
-    """
-    page_slug = 'test-slug'
-    page_content = 'test content'
-    page_title = 'test title'
-    page_isVisible = True
-
-    # test creating root page
-    variables = json.dumps({
-        'title': page_title, 'content': page_content,
-        'isVisible': page_isVisible, 'slug': page_slug})
-    response = admin_client.post(
-        reverse('dashboard:api'), {'query': query, 'variables': variables})
-    content = get_graphql_content(response)
-    assert 'errors' not in content
-    data = content['data']['pageCreate']
-    assert data['errors'] == []
-    assert data['page']['title'] == page_title
-    assert data['page']['content'] == page_content
-    assert data['page']['slug'] == page_slug
-    assert data['page']['isVisible'] == page_isVisible
-
-
-def test_page_delete_mutation(admin_client, page):
-    query = """
-        mutation DeletePage($id: ID!) {
-            pageDelete(id: $id) {
-                page {
-                    title
-                    id
-                }
-                errors {
-                    field
-                    message
-                }
-              }
-            }
-    """
-    variables = json.dumps({
-        'id': graphene.Node.to_global_id('Page', page.id)})
-    response = admin_client.post(
-        reverse('dashboard:api'), {'query': query, 'variables': variables})
-    content = get_graphql_content(response)
-    assert 'errors' not in content
-    data = content['data']['pageDelete']
-    assert data['page']['title'] == page.title
-    with pytest.raises(page._meta.model.DoesNotExist):
-        page.refresh_from_db()
